+++ src/scenes/CardsScene.ts
@@ -1,11 +1,13 @@
 import * as Phaser from 'phaser';
 import anvilImage from '../assets/anvil.png';
-import cardImage from '../assets/card.png';
+import cardDarkImage from '../assets/card-dark.png';
+import cardLightImage from '../assets/card-light.png';
 import deckImage from '../assets/deck.png';
 import ironImage from '../assets/resources/iron.png';
 import poisonImage from '../assets/resources/poison.png';
 import waterImage from '../assets/resources/water.png';
 import woodImage from '../assets/resources/wood.png';
+import weaponBulletImage from '../assets/weapon-bullet.png';
 import {Deck} from '../Deck';
 import {Hand} from '../Hand';
 import {constants} from '../utils/constants';
@@ -29,7 +31,9 @@ export class CardsScene extends Phaser.Scene {
     this.load.image('resource-water', waterImage);
     this.load.image('resource-iron', ironImage);
     this.load.image('resource-poison', poisonImage);
-    this.load.image('card', cardImage);
+    this.load.image('card-bullet', weaponBulletImage);
+    this.load.image('darkCard', cardDarkImage);
+    this.load.image('lightCard', cardLightImage);
   }
 
   create(): void {