+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Weapon swapping
+- blueprint and resource cards
 
 ## [0.0.5] - 2020-05-06
 ### Added